% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean the pheontype and expression data from TCGA}
\usage{
clean(phe, expr)
}
\arguments{
\item{phe}{data.frame, the raw expression matrix with colname of molecular id and individual ids.}

\item{expr}{data.frame, the raw expression matrix with colname of molecular id and individual ids.}
}
\value{
data.frame, the clean expression matrix with colname of molecular id and individual ids.
}
\description{
Delete the duplicated individual and rename the colnames.
}
\details{
nothing
}
\examples{
data(phe);data(expr)
cleanData = clean(phe,expr)

}
