% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleSummary.R
\name{simpleSummary}
\alias{simpleSummary}
\alias{simpleSummary.default}
\alias{simpleSummary.summary.coxph}
\title{simpleSummary}
\usage{
simpleSummary(summaryObj)

\method{simpleSummary}{default}(summaryObj)

\method{simpleSummary}{summary.coxph}(summaryObj)
}
\arguments{
\item{summaryObj}{The summary class}
}
\value{
a data.frame
}
\description{
This function would extract tidy information
}
